create progressBar component
